Time: 253 ms (13.77%), Space: 155.8 MB (10.64%) - LeetHub

+++ 2762-continuous-subarrays/2762-continuous-subarrays.cpp
@@ -3,7 +3,7 @@ class Solution {
 public:
     long long continuousSubarrays(vector<int>& a) {
         ll n = a.size(), ans = 0;
-        multiset<ll> s;
+        multiset<int> s;
         for (int i=0,j=0; j<n; j++) {
             s.insert(a[j]);
             while (s.size()) {
